Change route to app/dashboard
