+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/doc/en/button.md
@@ -1,0 +1,9 @@
+```
+<Button
+    className="bg-gradient-to-tr from-pink-500 to-yellow-500 text-white shadow-lg"
+    radius="full"
+    onPress={() => signout()}
+>
+    Sign Out
+</Button>
+```
