+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/app/[menu]/error.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { Button } from '@nextui-org/react'
 import { useEffect } from 'react'
 
 /**
@@ -27,14 +28,14 @@ export default function Error({
   return (
     <div>
       <h2>Something went wrong!</h2>
-      <button
+      <Button
         onClick={
           // Attempt to recover by trying to re-render the segment
           () => reset()
         }
       >
         Try again
-      </button>
+      </Button>
     </div>
   )
 }
