+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/app/[menu]/layout.tsx
@@ -1,0 +1,16 @@
+import SiteNavbar from '@/components/layout/SiteNavbar'
+import SiteFooter from '@/components/layout/SiteFooter'
+
+export default function Layout({
+  children,
+}: Readonly<{ children: React.ReactNode }>) {
+  return (
+    <div className="flex flex-col min-h-screen">
+      <SiteNavbar loggedIn={true} />
+      <main className="">{children}</main>
+      <footer className="mt-auto">
+        <SiteFooter />
+      </footer>
+    </div>
+  )
+}
