+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/app/[menu]/page.tsx
@@ -5,13 +5,13 @@ import { useSearchParams } from 'next/navigation'
 import DashboardManager from '@/feature/dashboard/DashboardManager'
 import AppContainer from '@/components/layout/AppContainer'
 
-export default function Page() {
+export default function Page({ params }: { params: { menu: string } }) {
   let searchParams = useSearchParams()
-  let currentSubMenu = searchParams.get('navigationKey') || ''
+  let currentSubMenu = searchParams.get('q') || '1efa02'
   let data = undefined
 
   return (
-    <AppContainer selectedMenu="dashboard" selectedSubMenu={currentSubMenu}>
+    <AppContainer selectedMenu={params.menu} selectedSubMenu={currentSubMenu}>
       <DashboardManager
         data={data ? data : [{}]}
         selectedSubMenu={currentSubMenu}
