--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/dashboard/layout.tsx
@@ -1,20 +1,0 @@
-import SiteNavigationBar from '@/components/layout/navigation/SiteNavigationBar'
-
-let data = {}
-
-let context = {
-  auth: undefined!,
-  navigationKey: 'example 01',
-  queryKey: '1efa02',
-}
-
-export default function Layout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <div>
-      <SiteNavigationBar siteSelector={true} />
-      <div>{children}</div>
-    </div>
-  )
-}
