+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/layout.tsx
@@ -1,14 +1,12 @@
+import './globals.css'
 import type { Metadata, Viewport } from 'next'
 
-import { AppNextUIProvider } from '@/utility/providers'
-
-import './globals.css'
 import React from 'react'
+import clsx from 'clsx'
 
+import { AppNextUIProvider } from '@/utility/providers'
 import { siteConfig } from '@/utility/constants'
 import { fontSans } from '@/utility/fonts'
-
-import clsx from 'clsx'
 
 export const metadata: Metadata = {
   title: {
@@ -25,17 +23,12 @@ export const viewport: Viewport = {
   ],
 }
 
-/**
- * Root layout component.
- *
- * @param children - The children components to be rendered.
- * @returns The rendered RootLayout component.
- */
 export default function RootLayout({
   children,
 }: Readonly<{ children: React.ReactNode }>) {
   return (
     <html suppressHydrationWarning lang="en">
+      {/* Suppress hyrdration warning was taken from nextui new project template. */}
       <head />
       <body
         className={clsx(
@@ -47,7 +40,7 @@ export default function RootLayout({
         <AppNextUIProvider
           themeProps={{ attribute: 'class', children, defaultTheme: 'light' }}
         >
-          <div>{children}</div>
+          {children}
         </AppNextUIProvider>
       </body>
     </html>
