+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/page.tsx
@@ -1,12 +1,18 @@
-import SiteNavigationBar from '@/components/layout/navigation/SiteNavigationBar'
+import SiteNavbar from '@/components/layout/SiteNavbar'
+import SiteFooter from '@/components/layout/SiteFooter'
 import LandingPage02 from '@/feature/landingPage/LandingPage02'
 
 export default async function Page() {
   return (
     <div className="h-screen v-screen">
-      <SiteNavigationBar siteSelector={false} />
-      <div className="bg-blue-100">
-        <LandingPage02 />
+      <div className="flex flex-col min-h-screen">
+        <SiteNavbar loggedIn={false} />
+        <main>
+          <LandingPage02 />
+        </main>
+        <footer className="mt-auto">
+          <SiteFooter />
+        </footer>
       </div>
     </div>
   )
