--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/app/student/loading.tsx
@@ -1,5 +1,0 @@
-import LoadingSkeleton from '@/components/ui/skeleton/LoadingSkeleton'
-
-export default function Loader() {
-  return <LoadingSkeleton />
-}
