+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/layout/AppContainer.tsx
@@ -1,4 +1,4 @@
-import AppNavigationBar from '@/components/layout/navigation/AppNavigationBar'
+import AppNavigationBar from '@/components/layout/AppNavigationBar'
 
 interface AppContainerProps {
   selectedMenu: string
