+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/layout/SiteFooter.tsx
@@ -1,17 +1,18 @@
-function Footer() {
+import { Navbar, NavbarContent, NavbarItem } from '@nextui-org/react'
+
+export function SiteFooter() {
   return (
-    <div>
-      <div>
-        <span
-          style={{ float: 'left', paddingRight: '16px', fontWeight: '600' }}
-        >
-          Terms of Use
-        </span>
-        <span style={{ float: 'left', fontWeight: '600' }}>Privacy Policy</span>
-        <span style={{ float: 'right', fontWeight: '600' }}>Disclaimer</span>
-      </div>
-    </div>
+    <Navbar className="font-semibold">
+      <NavbarContent justify="start">
+        <NavbarItem>Terms of Use</NavbarItem>
+        <NavbarItem>Privacy Policy</NavbarItem>
+        <NavbarItem>HIPPA</NavbarItem>
+      </NavbarContent>
+      <NavbarContent justify="end">
+        <NavbarItem>2024 © Open Sacramento</NavbarItem>
+      </NavbarContent>
+    </Navbar>
   )
 }
 
-export default Footer
+export default SiteFooter
