--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/ui/card/old/CardGroupOld.module.css
@@ -1,66 +1,0 @@
-.container {
-  display: grid;
-  grid-template-columns: repeat(1, 1fr);
-  gap: 10px;
-  grid-auto-rows: minmax(100px, auto);
-  width: 100%;
-}
-
-.cardBodyImageContainer {
-  display: flex;
-  justify-content: center;
-}
-
-@media (min-width: 400px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
-
-@media (min-width: 576px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
-
-@media (min-width: 768px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
-
-@media (min-width: 992px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
-
-@media (min-width: 1200px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
-
-@media (min-width: 1400px) {
-  .container {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 10px;
-    grid-auto-rows: minmax(100px, auto);
-  }
-}
