--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/ui/card/old/HomeCards/Card02.tsx
@@ -1,28 +1,0 @@
-import Card from 'react-bootstrap/Card'
-import ListGroup from 'react-bootstrap/ListGroup'
-
-function Card02() {
-  return (
-    <Card style={{ width: '18rem' }}>
-      <Card.Img src="/demo_04.png" variant="top" />
-      <Card.Body>
-        <Card.Title>Bar Graphs from data</Card.Title>
-        <Card.Text>
-          Some quick example text to build on the card title and make up the
-          bulk of the card's content.
-        </Card.Text>
-      </Card.Body>
-      <ListGroup className="list-group-flush">
-        <ListGroup.Item>Cras justo odio</ListGroup.Item>
-        <ListGroup.Item>Dapibus ac facilisis in</ListGroup.Item>
-        <ListGroup.Item>Vestibulum at eros</ListGroup.Item>
-      </ListGroup>
-      <Card.Body>
-        <Card.Link href="#">Card Link</Card.Link>
-        <Card.Link href="#">Another Link</Card.Link>
-      </Card.Body>
-    </Card>
-  )
-}
-
-export default Card02
