--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/ui/card/old/HomeCards/Form01.tsx
@@ -1,108 +1,0 @@
-import type { FieldApi } from '@tanstack/react-form'
-
-import { useForm } from '@tanstack/react-form'
-import { Button } from 'react-bootstrap'
-
-function FieldInfo({ field }: { field: FieldApi<any, any, any, any> }) {
-  return (
-    <>
-      {field.state.meta.touchedErrors ? (
-        <em>{field.state.meta.touchedErrors} </em>
-      ) : null}
-      {field.state.meta.isValidating ? 'Validating...' : null}
-    </>
-  )
-}
-
-function Form01() {
-  const form = useForm({
-    defaultValues: {
-      firstName: '',
-      lastName: '',
-    },
-    onSubmit: async ({ value }) => {
-      // Do something with form data
-      console.log(value)
-    },
-  })
-
-  return (
-    <div>
-      <h1>Simple Form Example </h1>
-      <form.Provider>
-        <form
-          onSubmit={(e) => {
-            e.preventDefault()
-            e.stopPropagation()
-            void form.handleSubmit()
-          }}
-        >
-          <div>
-            {/* A type-safe field component*/}
-            <form.Field
-              children={(field) => {
-                // Avoid hasty abstractions. Render props are great!
-                return (
-                  <>
-                    <label htmlFor={field.name}> First Name: </label>
-                    <input
-                      name={field.name}
-                      value={field.state.value}
-                      onBlur={field.handleBlur}
-                      onChange={(e) => field.handleChange(e.target.value)}
-                    />
-                    <FieldInfo field={field} />
-                  </>
-                )
-              }}
-              name="firstName"
-              validators={{
-                onChange: ({ value }) =>
-                  !value
-                    ? 'A first name is required'
-                    : value.length < 3
-                      ? 'First name must be at least 3 characters'
-                      : undefined,
-                onChangeAsyncDebounceMs: 500,
-                onChangeAsync: async ({ value }) => {
-                  await new Promise((resolve) => setTimeout(resolve, 1000))
-                  return (
-                    value.includes('error') &&
-                    'No "error" allowed in first name'
-                  )
-                },
-              }}
-            />
-          </div>
-          <div>
-            <form.Field
-              children={(field) => (
-                <>
-                  <label htmlFor={field.name}> Last Name: </label>
-                  <input
-                    name={field.name}
-                    value={field.state.value}
-                    onBlur={field.handleBlur}
-                    onChange={(e) => field.handleChange(e.target.value)}
-                  />
-                  <FieldInfo field={field} />
-                </>
-              )}
-              name="lastName"
-            />
-          </div>
-          <form.Subscribe
-            children={([canSubmit, isSubmitting]) => (
-              <Button disabled={!canSubmit} type="submit">
-                {isSubmitting ? '...' : 'Submit'}
-              </Button>
-            )}
-            selector={(state) => [state.canSubmit, state.isSubmitting]}
-          />
-        </form>
-      </form.Provider>
-    </div>
-  )
-}
-
-export default Form01
