--- LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/components/ui/card/old/HomeCards/HomeDashboardGrid.module.css
@@ -1,9 +1,0 @@
-.dashboard_header {
-  /* Heading/H1 - Small/Semi Bold */
-  font-size: 39px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: 47px; /* 120.513% */
-  letter-spacing: -1.56px;
-  padding: 16px 0 24px 0;
-}
