+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/feature/dashboard/DashboardManager.tsx
@@ -1,6 +1,7 @@
-import DashboardD81 from '@/feature/dashboard/DashboardD81'
+import React from 'react'
 
 export interface DashboardManagerProps {
+  children: React.ReactNode
   data: [any]
   selectedSubMenu: string
   setPageQueryKey?: any
@@ -10,18 +11,10 @@ export interface DashboardManagerProps {
  * DashboardManager component to manage and render different dashboard views
  * based on the selected submenu.
  *
- * @param {DashboardManagerProps} props - The props for the DashboardManager component.
- * @returns {ReactElement} The rendered dashboard component.
- */
-export default function DashboardManager({
   data,
   selectedSubMenu,
   setPageQueryKey,
-}: DashboardManagerProps) {
-  return (
-    <div>
-      <DashboardD81 data={data} setPageQueryKey={setPageQueryKey} />
-      {/* <DashboardD0f demographics={data}></DashboardD0f> */}
-    </div>
-  )
+ */
+export default function DashboardManager({ children }: DashboardManagerProps) {
+  return <div>{children}</div>
 }
