+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/src/utility/providers.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import { NextUIProvider } from '@nextui-org/system'
 import * as React from 'react'
 import { ThemeProviderProps } from 'next-themes/dist/types'
