+++ LB_Community_Edition_Non_Enterprise/Pilot_Iteration_Files/gcc-iteration-nextjs/tailwind.config.ts
@@ -21,7 +21,36 @@ const config: Config = {
     },
   },
   darkMode: 'class',
-  plugins: [nextui({ prefix: 'lb' })],
+  plugins: [
+    nextui({
+      prefix: 'lb',
+      layout: {
+        disabledOpacity: '0.3', // opacity-[0.3]
+        radius: {
+          small: '2px', // rounded-small
+          medium: '4px', // rounded-medium
+          large: '6px', // rounded-large
+        },
+        borderWidth: {
+          small: '1px', // border-small
+          medium: '1px', // border-medium
+          large: '2px', // border-large
+        },
+      },
+      themes: {
+        light: {
+          colors: {
+            secondary: {
+              DEFAULT: '#003049',
+              foreground: '#000000',
+            },
+            focus: '#003049',
+          },
+        },
+        dark: {},
+      },
+    }),
+  ],
 }
 
 export default config
